backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
transformers==4.47.1
torch>=2.6.0,<3.0
protobuf==5.29.2
python-dotenv==1.0.1
pydantic==2.10.4
pytest==8.3.4
httpx==0.28.1
